fixed issu in repository class

+++ hibernate_OneToOne/src/main/java/lk/harshana/dao/AccountDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import lk.harshana.model.Account;
 
-public interface AccountDao extends CrudRepository<Account, Long>{
+public interface AccountDao extends CrudRepository<Account, Integer>{
 
 }
